Fixed unnecessary step in docker_stack_deploy.sh listing.
Traefik does not need to be restarted at all or have the main network removed to bounce any stack. In fact, that's why the instructions have you build that network first manually, rather than specify it in the traefik.yml. I bounce traefik by itself all the time but leave its network alone. That is key.